requirements.txt
# Autocaption Web - Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
openai-whisper==20231117
ffmpeg-python==0.2.0
aiofiles==23.2.1
pyngrok==7.0.11
